Update for login and registration method
