--- mmdt/accounts/forms.py
@@ -1,8 +1,0 @@
-from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth.models import User
-
-class CustomUserCreationForm(UserCreationForm):
-    class Meta:
-        model = User
-        fields = ['username', 'email', 'password1', 'password2']
-
