--- mmdt/accounts/views.py
@@ -1,15 +1,0 @@
-from django.contrib.auth import login
-from django.shortcuts import render, redirect
-from .forms import CustomUserCreationForm
-
-def register(request):
-    if request.method == 'POST':
-        form = CustomUserCreationForm(request.POST)
-        if form.is_valid():
-            user = form.save()
-            login(request, user)
-            return redirect('home')
-    else:
-        form = CustomUserCreationForm()
-    
-    return render(request, 'accounts/register.html', {'form': form})
