+++ mmdt/mmdt/urls.py
@@ -24,11 +24,11 @@ from django.contrib.auth.views import LoginView, LogoutView
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('blog.urls')),
+    path('', include('users.urls', namespace='users')),
     path('polls/', include('polls.urls', namespace='polls')),
     path('survey/', include('survey.urls', namespace='survey')),
-    path('accounts/', include('accounts.urls', namespace='accounts')),
     path('summernote/', include('django_summernote.urls')),
-    path('accounts/login/', LoginView.as_view(template_name='accounts/login.html'), name='login'),
+    path('accounts/login/', LoginView.as_view(template_name='registration/login.html'), name='login'),
     path('accounts/logout/', LogoutView.as_view(next_page='/'), name='logout'),
 ]
