+++ mmdt/polls/urls.py
@@ -7,7 +7,6 @@ app_name = "polls"
 urlpatterns = [
     path("", views.PollHomePage.index, name="index"),
     path('vote/', views.PollHomePage.vote, name='vote'),
-    path('register/', views.register, name='register'),
     path('login/', LoginView.as_view(template_name='polls/login.html'), name='login'),
     path('logout/', LogoutView.as_view(next_page='/'), name='logout'),
     path('all_results/', views.PollHomePage.all_results, name='all_results'),
