+++ mmdt/users/forms.py
@@ -1,0 +1,7 @@
+# users/forms.py
+
+from django.contrib.auth.forms import UserCreationForm
+
+class CustomUserCreationForm(UserCreationForm):
+    class Meta(UserCreationForm.Meta):
+        fields = UserCreationForm.Meta.fields + ("email",)
